+++ packages/bluesky/lib/src/services/codegen/app/bsky/notification/listNotifications/output.g.dart
@@ -25,7 +25,6 @@ _$NotificationListNotificationsOutputFromJson(Map json) => $checkedCreate(
             )
             .toList(),
       ),
-      priority: $checkedConvert('priority', (v) => v as bool?),
       seenAt: $checkedConvert(
         'seenAt',
         (v) => v == null ? null : DateTime.parse(v as String),
@@ -46,7 +45,6 @@ Map<String, dynamic> _$NotificationListNotificationsOutputToJson(
   'notifications': instance.notifications
       .map(const NotificationConverter().toJson)
       .toList(),
-  'priority': ?instance.priority,
   'seenAt': iso8601(instance.seenAt),
   r'$unknown': ?instance.$unknown,
 };