+++ packages/bluesky/lib/src/services/codegen/app/bsky/notification_service.dart
@@ -64,7 +64,6 @@ appBskyNotificationGetPreferences({
 /// Count the number of unread notifications for the requesting account. Requires auth.
 Future<XRPCResponse<NotificationGetUnreadCountOutput>>
 appBskyNotificationGetUnreadCount({
-  bool? priority,
   DateTime? seenAt,
   required ServiceContext $ctx,
   String? $service,
@@ -74,11 +73,7 @@ appBskyNotificationGetUnreadCount({
   ns.appBskyNotificationGetUnreadCount,
   service: $service,
   headers: $headers,
-  parameters: {
-    ...?$unknown,
-    if (priority != null) 'priority': priority,
-    if (seenAt != null) 'seenAt': iso8601(seenAt),
-  },
+  parameters: {...?$unknown, if (seenAt != null) 'seenAt': iso8601(seenAt)},
   to: const NotificationGetUnreadCountOutputConverter().fromJson,
 );
 
@@ -108,7 +103,6 @@ Future<XRPCResponse<NotificationListNotificationsOutput>>
 appBskyNotificationListNotifications({
   List<String>? reasons,
   int? limit,
-  bool? priority,
   String? cursor,
   DateTime? seenAt,
   required ServiceContext $ctx,
@@ -123,7 +117,6 @@ appBskyNotificationListNotifications({
     ...?$unknown,
     if (reasons != null) 'reasons': reasons,
     if (limit != null) 'limit': limit,
-    if (priority != null) 'priority': priority,
     if (cursor != null) 'cursor': cursor,
     if (seenAt != null) 'seenAt': iso8601(seenAt),
   },
@@ -317,13 +310,11 @@ base class NotificationService {
 
   /// Count the number of unread notifications for the requesting account. Requires auth.
   Future<XRPCResponse<NotificationGetUnreadCountOutput>> getUnreadCount({
-    bool? priority,
     DateTime? seenAt,
     String? $service,
     Map<String, String>? $headers,
     Map<String, String>? $unknown,
   }) async => await appBskyNotificationGetUnreadCount(
-    priority: priority,
     seenAt: seenAt,
     $ctx: ctx,
     $service: $service,
@@ -352,7 +343,6 @@ base class NotificationService {
   Future<XRPCResponse<NotificationListNotificationsOutput>> listNotifications({
     List<String>? reasons,
     int? limit,
-    bool? priority,
     String? cursor,
     DateTime? seenAt,
     String? $service,
@@ -361,7 +351,6 @@ base class NotificationService {
   }) async => await appBskyNotificationListNotifications(
     reasons: reasons,
     limit: limit,
-    priority: priority,
     cursor: cursor,
     seenAt: seenAt,
     $ctx: ctx,